+++ orchestrator/priv/repo/migrations/20241223013000_add_hmac_secret_to_push_configs.exs
@@ -1,9 +1,0 @@
-defmodule Orchestrator.Repo.Migrations.AddHmacSecretToPushConfigs do
-  use Ecto.Migration
-
-  def change do
-    alter table(:push_configs) do
-      add :hmac_secret, :string
-    end
-  end
-end